Se agrega la funcion para insertar servicios y tecnicos

+++ DB/tablas.sql
@@ -12,7 +12,7 @@ CREATE TABLE servicios(
     nombre VARCHAR(100) NOT NULL,
     clave VARCHAR(50) NOT NULL,
     precio FLOAT(6,2),
-    duracion_hrs INT(11)
+    duracion_hrs FLOAT(6, 2)
 );
 
 CREATE TABLE pedidos(
